+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "single-scripture-rcl",
   "description": "A React Component Library for Rendering a Single Scripture Resource.",
-  "version": "0.1.8",
+  "version": "0.1.6",
   "repository": {
     "type": "git",
     "url": "https://github.com/unfoldingWord/single-scripture-rcl.git"
@@ -36,25 +36,20 @@
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@typescript-eslint/eslint-plugin": "^4.14.1",
     "@unfoldingword/eslint-config": "1.3.0",
     "babel-loader": "^8.1.0",
     "core-js": "3.7.0",
     "css-loader": "^5.0.1",
-    "eslint": "^7.16.0",
+    "eslint": "7.13.0",
     "eslint-config-prettier": "6.15.0",
-    "eslint-config-react-app": "^6.0.0",
     "eslint-config-recommended": "4.1.0",
     "eslint-plugin-chai-friendly": "0.6.0",
     "eslint-plugin-cypress": "2.11.2",
-    "eslint-plugin-flowtype": "^5.2.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jest": "24.1.3",
-    "eslint-plugin-jsdoc": "^31.4.0",
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-react": "7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
     "jest": "^26.6.3",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
