--- src/components/ComboBox/index.ts
@@ -1,1 +1,0 @@
-export { ComboBox } from './ComboBox'
