--- src/components/ScriptureCard/index.ts
@@ -1,1 +1,0 @@
-export { default as ScriptureCard } from './ScriptureCard';
