--- src/components/ScriptureSelector/index.ts
@@ -1,1 +1,0 @@
-export { default as ScriptureSelector } from './ScriptureSelector'
