+++ src/components/index.ts
@@ -1,4 +1,1 @@
-export * from './ComboBox'
-export * from './ScriptureCard'
-export * from './ScripturePane'
-export * from './ScriptureSelector'
+export * from './ScripturePane';
