+++ src/index.ts
@@ -1,11 +1,2 @@
-export * from './components'
-export * from './utils/common'
-export {
-  useComboBox,
-  useResourceManifest,
-  useScripture,
-  useScriptureResources,
-  useScriptureSelector,
-  useScriptureSettings,
-} from './hooks'
-
+export * from './components';
+export { useScripture, useResourceManifest } from './hooks';
