--- src/utils/common.ts
@@ -1,7 +1,0 @@
-export const NT_ORIG_LANG = 'el-x-koine'
-export const NT_ORIG_LANG_BIBLE = 'ugnt'
-export const OT_ORIG_LANG = 'hbo'
-export const OT_ORIG_LANG_BIBLE = 'uhb'
-export const ORIGINAL_SOURCE = 'ORIGINAL_SOURCE'
-export const TARGET_LITERAL = 'TARGET_LITERAL'
-export const TARGET_SIMPLIFIED = 'TARGET_SIMPLIFIED'
