--- src/utils/delay.ts
@@ -1,5 +1,0 @@
-export function delay(ms) {
-  return new Promise((resolve) =>
-    setTimeout(resolve, ms),
-  )
-}
