+++ styleguide.config.js
@@ -12,50 +12,8 @@ let sections = [
     content: 'README.md',
   },
   {
-    name: 'ScriptureCard ',
-    content: 'src/components/ScriptureCard/README.md',
-    components: () => {
-      const componentNames = ['Resource.context'];
-      return componentNames.map((componentName) => {
-        return path.resolve(
-          __dirname,
-          `src/components/resources`,
-          `${componentName}.js`,
-        );
-      });
-    },
-  },
-  {
     name: 'ScripturePane ',
     content: 'src/components/ScripturePane/README.md',
-    components: () => {
-      const componentNames = ['Resource.context'];
-      return componentNames.map((componentName) => {
-        return path.resolve(
-          __dirname,
-          `src/components/resources`,
-          `${componentName}.js`,
-        );
-      });
-    },
-  },
-  {
-    name: 'ScriptureSelector ',
-    content: 'src/components/ScriptureSelector/README.md',
-    components: () => {
-      const componentNames = ['Resource.context'];
-      return componentNames.map((componentName) => {
-        return path.resolve(
-          __dirname,
-          `src/components/resources`,
-          `${componentName}.js`,
-        );
-      });
-    },
-  },
-  {
-    name: 'ComboBox ',
-    content: 'src/components/ComboBox/README.md',
     components: () => {
       const componentNames = ['Resource.context'];
       return componentNames.map((componentName) => {
